springboot 集成 shardingProxy 案例

+++ spring-lab/lab-sharding-datasource/lab-sharding-datasource-01/src/main/java/org/springframework/springboot/lab/sharding/mapper/OrderMapper.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface OrderMapper {
 
-    OrderDO selectById(@Param("id") Integer id);
+    OrderDO selectById(@Param("id") Long id);
 
     List<OrderDO> selectListByUserId(@Param("userId") Integer userId);
